Add BidBotEcho to nick_aliases.txt
Admittedly, this is only relevant for a short time during the year, but
still.  I'm not sure if it should be treated as an alias to bidserv, if
bidserv should be an alias to BidBotEcho, or if BidBotEcho should be its
own entry, but at any rate, it should be treated as a bot, not a user.